continue button component
